Move Main to separate file
